requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
crispy-bootstrap5==2024.2
cryptography==43.0.1
daphne==4.1.2
Django==5.1
django-appconf==1.0.6
django-autoslug==1.9.9
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-imagekit==5.0.0
django-phonenumber-field==8.0.0
django-tables2==2.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
openpyxl==3.1.5
phonenumbers==8.13.43
pilkit==3.0
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
requests==2.32.3
rsa==4.9
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tablib==3.6.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
zope.interface==7.0.3
